program2.java
This Java program checks whether a given number is even or odd. It contains two classes: Logic and Program2. The Logic class has a method named CeckEvenOdd(int iNo) which accepts an integer as an argument. Inside this method, an if-else condition is used to determine the parity of the number. The condition iNo % 2 == 0 checks whether the number is divisible by 2. If the remainder is zero, the method prints "Number is Even"; otherwise, it prints "Number is Odd". The Program2 class contains the main method, where an object of the Logic class is created. Using this object, the CeckEvenOdd() method is called with the value 7. The program then checks the number and displays that 7 is Odd. This example demonstrates the use of conditional statements and object-oriented method calls in Java.

Signed-off-by: Apurva Satish Ratnakar <apurvasatishratnakar1409@gmail.com>